cut/paste dump* function implementations from COOLWSD.cpp into dumpWsdState.cpp
