+++ android/lib/src/main/cpp/CMakeLists.txt.in
@@ -38,6 +38,7 @@ add_library(androidapp SHARED
             ../../../../../wsd/ClientSession.cpp
             ../../../../../wsd/DocumentBroker.cpp
             ../../../../../wsd/COOLWSD.cpp
+            ../../../../../wsd/dumpWsdState.cpp
             ../../../../../wsd/ClientRequestDispatcher.cpp
             ../../../../../wsd/RequestDetails.cpp
             ../../../../../wsd/RequestVettingStation.cpp
