+++ gtk/Makefile.am
@@ -59,6 +59,7 @@ cmake_list= \
             ../wsd/ClientSession.cpp \
             ../wsd/DocumentBroker.cpp \
             ../wsd/COOLWSD.cpp \
+            ../wsd/dumpWsdState.cpp \
             ../wsd/ClientRequestDispatcher.cpp \
             ../wsd/RequestDetails.cpp \
             ../wsd/RequestVettingStation.cpp \
