+++ wsd/COOLWSD.cpp
@@ -2264,60 +2264,6 @@ void COOLWSD::initializeSSL()
 #endif
 }
 
-void COOLWSD::dumpNewSessionTrace(const std::string& id, const std::string& sessionId, const std::string& uri, const std::string& path)
-{
-    if (TraceDumper)
-    {
-        try
-        {
-            TraceDumper->newSession(id, sessionId, uri, path);
-        }
-        catch (const std::exception& exc)
-        {
-            LOG_ERR("Exception in tracer newSession: " << exc.what());
-        }
-    }
-}
-
-void COOLWSD::dumpEndSessionTrace(const std::string& id, const std::string& sessionId, const std::string& uri)
-{
-    if (TraceDumper)
-    {
-        try
-        {
-            TraceDumper->endSession(id, sessionId, uri);
-        }
-        catch (const std::exception& exc)
-        {
-            LOG_ERR("Exception in tracer newSession: " << exc.what());
-        }
-    }
-}
-
-void COOLWSD::dumpEventTrace(const std::string& id, const std::string& sessionId, const std::string& data)
-{
-    if (TraceDumper)
-    {
-        TraceDumper->writeEvent(id, sessionId, data);
-    }
-}
-
-void COOLWSD::dumpIncomingTrace(const std::string& id, const std::string& sessionId, const std::string& data)
-{
-    if (TraceDumper)
-    {
-        TraceDumper->writeIncoming(id, sessionId, data);
-    }
-}
-
-void COOLWSD::dumpOutgoingTrace(const std::string& id, const std::string& sessionId, const std::string& data)
-{
-    if (TraceDumper)
-    {
-        TraceDumper->writeOutgoing(id, sessionId, data);
-    }
-}
-
 void COOLWSD::defineOptions(Poco::Util::OptionSet& optionSet)
 {
     if constexpr (Util::isMobileApp())
