+++ wsd/dumpWsdState.cpp
@@ -1,0 +1,73 @@
+/* -*- Mode: C++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4; fill-column: 100 -*- */
+/*
+ * Copyright the Collabora Online contributors.
+ *
+ * SPDX-License-Identifier: MPL-2.0
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+#include <config.h>
+
+#include <wsd/COOLWSD.hpp>
+#include <wsd/TraceFile.hpp>
+
+#include <string>
+
+void COOLWSD::dumpNewSessionTrace(const std::string& id, const std::string& sessionId, const std::string& uri, const std::string& path)
+{
+    if (TraceDumper)
+    {
+        try
+        {
+            TraceDumper->newSession(id, sessionId, uri, path);
+        }
+        catch (const std::exception& exc)
+        {
+            LOG_ERR("Exception in tracer newSession: " << exc.what());
+        }
+    }
+}
+
+void COOLWSD::dumpEndSessionTrace(const std::string& id, const std::string& sessionId, const std::string& uri)
+{
+    if (TraceDumper)
+    {
+        try
+        {
+            TraceDumper->endSession(id, sessionId, uri);
+        }
+        catch (const std::exception& exc)
+        {
+            LOG_ERR("Exception in tracer newSession: " << exc.what());
+        }
+    }
+}
+
+void COOLWSD::dumpEventTrace(const std::string& id, const std::string& sessionId, const std::string& data)
+{
+    if (TraceDumper)
+    {
+        TraceDumper->writeEvent(id, sessionId, data);
+    }
+}
+
+void COOLWSD::dumpIncomingTrace(const std::string& id, const std::string& sessionId, const std::string& data)
+{
+    if (TraceDumper)
+    {
+        TraceDumper->writeIncoming(id, sessionId, data);
+    }
+}
+
+void COOLWSD::dumpOutgoingTrace(const std::string& id, const std::string& sessionId, const std::string& data)
+{
+    if (TraceDumper)
+    {
+        TraceDumper->writeOutgoing(id, sessionId, data);
+    }
+}
+
+/* vim:set shiftwidth=4 softtabstop=4 expandtab: */
